For time range, time will be in milliseconds rather than seconds. Do necessary conversions.
git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtWebApp/branches/V2transcript@21449 3dfb626b-079f-4ab1-a9f6-4654e5377508

+++ src/main/javascript/NextThought/model/anchorables/TranscriptContentPointer.js
@@ -16,7 +16,7 @@ Ext.define('NextThought.model.anchorables.TranscriptContentPointer', {
 				pointer: cp.createFromObject(o.pointer),
 				cueid: o.cueid,
 				role: o.role,
-				seconds: parseFloat(o.seconds)
+				seconds: parseInt(o.seconds)
 			});
 		}
 	},
